refactor to follow Python conventions and consistency - file renaming
