+++ sentence_classification/automatic_learning.py
@@ -1,5 +1,5 @@
 import sys, numpy as np
-from dataUtils import *
+from data_utils import *
 import pipelines
 from random import shuffle
